X 型键盘: 支持按照普通的按键尺寸设置中心按键尺寸并根据实际可用空间缩放 X 面板

+++ app/src/main/java/org/crazydan/studio/app/ime/kuaizi/internal/keyboard/state/EditorEditDoingStateData.java
@@ -31,7 +31,7 @@ public class EditorEditDoingStateData implements State.Data {
 
     public static Motion createAnchor(Motion motion) {
         // 根据屏幕移动距离得出光标移动字符数
-        int distance = motion.distance > 0 ? Math.max(1, motion.distance / ScreenUtils.dpToPx(16)) : 0;
+        float distance = motion.distance > 0 ? Math.max(1, motion.distance / ScreenUtils.dpToPx(16f)) : 0;
 
         return new Motion(motion.direction, distance, motion.timestamp);
     }
